completed the home page as well as enabled logout

+++ api/v1/views/votes.py
@@ -4,9 +4,11 @@ handling of the votes api view
 """
 from . import Post, Comment
 from . import Vote, User
-from flask import abort, jsonify
+from flask import abort, jsonify, current_app, g
 from . import app_views, storage
-from .secure import auth, verify_password
+with current_app.app_context():
+    from .secure import auth, verify_password
+
 
 @app_views.route("/posts/<an_id>/votes", strict_slashes=False)
 def article_votes(an_id):
@@ -75,8 +77,12 @@ def has_voted(user_id, an_id):
 def delete_vote(vote_id):
     """ deletes a vote instance"""
     vote = storage.get(Vote, vote_id)
-
     if vote is None:
         abort(404)
+
+    """ making sure only the owner of the vote is allowed"""
+    if g.user.name != vote.voter.name:
+        abort(401)
+
     storage.delete(vote)
     storage.save()
